break soldier jumping into three seperate functions

+++ level.c
@@ -61,7 +61,7 @@ u16 is_leftmost(void) {
     return (*back >> 8) == 0 && (window <= WINDOW_MIN);
 }
 
-byte platform_bottom(void) {
+u16 platform_bottom(void) {
     return platform_count > 0 ? height[1 + platform_count] : 0;
 }
 
@@ -131,9 +131,9 @@ void update_window(short direction) {
     }
 }
 
-byte get_snap(byte prev, byte next) {
+u16 get_snap(u16 prev, u16 next) {
     for (u16 i = 0; i < platform_count; i++) {
-	byte snap = height[HEIGHT_DATA + i];
+	u16 snap = height[HEIGHT_DATA + i];
 	if (prev <= snap && snap <= next) {
 	    return snap;
 	}
